Refactor aws_runners path to match module name
