+++ github-runner-provisioner/internal/aws/ec2.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/aws/aws-sdk-go-v2/service/ec2/types"
-	"github.com/datawire/infra-actions/github-runner-provisioner/internal/aws/runners"
+	"github.com/datawire/infra-actions/github-runner-provisioner/internal/aws/aws_runners"
 	"github.com/datawire/infra-actions/github-runner-provisioner/internal/utils"
 	log "github.com/sirupsen/logrus"
 	"time"
