+++ github-runner-provisioner/kustomization.yaml
@@ -21,9 +21,11 @@ configMapGenerator:
       - internal/aws/ec2.go
       - internal/aws/iface.go
       - internal/aws/runner.go
-      - internal/aws/runners/macm1.go
-      - internal/aws/runners/ubuntuarm64.go
-      - internal/aws/runners/shared.go
+  - name: internal-aws-runners
+    files:
+      - internal/aws/aws_runners/macm1.go
+      - internal/aws/aws_runners/ubuntuarm64.go
+      - internal/aws/aws_runners/shared.go
   - name: internal-codemagic
     files:
       - internal/codemagic/runner.go
